+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.2
+* bugfix in keyword extractor
+* bugfix polygonPart
+* make kwsExtractor observable
+
 ## 2.1
 * add optional time to KWS result
 * bugfix: create logical baselin-polygon from json-string on demand